brahma_hq_uploader/requirements.txt
fastapi
uvicorn
google-cloud-storage>=2.14.0
google-auth>=2.25.0
google-api-core>=2.15.0
python-multipart
